Renaming: SpecificProof -> ProofSkeleton, merge ClosedTheoryItem into ClosedTheory
