--- lib/theory/src/Items/ClosedTheoryItem.hs
@@ -1,35 +1,0 @@
-module Items.ClosedTheoryItem (
-    module Items.ClosedTheoryItem
-) where
-
-import           Prelude                             hiding (id, (.))
-
--- import qualified Data.Label.Total
-
-
-import           Theory.Model
-import           Theory.Proof
-
-
-import TheoryObject
-import Rule
-
--- | Closed theories can be proven. Invariants:
---     1. Lemma names are unique
---     2. All proof steps with annotated sequents are sound with respect to the
---        closed rule set of the theory.
---     3. Maude is running under the given handle.
-type ClosedTheory =
-    Theory SignatureWithMaude ClosedRuleCache ClosedProtoRule IncrementalProof ()
-
--- | Closed Diff theories can be proven. Invariants:
---     1. Lemma names are unique
---     2. All proof steps with annotated sequents are sound with respect to the
---        closed rule set of the theory.
---     3. Maude is running under the given handle.
-type ClosedDiffTheory =
-    DiffTheory SignatureWithMaude ClosedRuleCache DiffProtoRule ClosedProtoRule IncrementalDiffProof IncrementalProof
-
--- | Either Therories can be Either a normal or a diff theory
-
-type EitherClosedTheory = Either ClosedTheory ClosedDiffTheory
