+++ lib/theory/src/Items/OpenTheoryItem.hs
@@ -2,15 +2,11 @@ module Items.OpenTheoryItem (
     module Items.OpenTheoryItem
 ) where
 
-import           Prelude                             hiding (id, (.))
-
-import           Theory.Model
-
-import TheoryObject
-
-import SpecificProof
-
 import Rule
+import Theory.ProofSkeleton
+import Theory.Model
+import TheoryObject
+import Prelude hiding (id, (.))
 
 -- | Open theories can be extended. Invariants:
 --   1. Lemma names are unique.
@@ -18,7 +14,6 @@ type OpenTheory =
     Theory SignaturePure [IntrRuleAC] OpenProtoRule ProofSkeleton SapicElement
 
 type OpenTheoryItem = TheoryItem OpenProtoRule ProofSkeleton SapicElement
-
 
 -- | Open theories can be extended. Invariants:
 --   1. Lemma names are unique.
